♻️ CartItemCard로 분리
- getAppliedDiscount는 util함수로 분리
