+++ src/refactoring/components/CartPage.tsx
@@ -1,5 +1,6 @@
-import { CartItem, Coupon, Product } from '../../types.ts';
+import { Coupon, Product } from '../../types.ts';
 import { useCart } from '../hooks';
+import CartItemCard from './CartItemCard.tsx';
 
 interface Props {
   products: Product[];
@@ -17,9 +18,8 @@ export const CartPage = ({ products, coupons }: Props) => {
     selectedCoupon,
   } = useCart();
 
-  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
-    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-  };
+  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) =>
+    discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
 
   const getRemainingStock = (product: Product) => {
     const cartItem = cart.find((item) => item.product.id === product.id);
@@ -27,18 +27,6 @@ export const CartPage = ({ products, coupons }: Props) => {
   };
 
   const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } = calculateTotal();
-
-  const getAppliedDiscount = (item: CartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
 
   return (
     <div className='container mx-auto p-4'>
@@ -100,48 +88,13 @@ export const CartPage = ({ products, coupons }: Props) => {
           <h2 className='text-2xl font-semibold mb-4'>장바구니 내역</h2>
 
           <div className='space-y-2'>
-            {cart.map((item) => {
-              const appliedDiscount = getAppliedDiscount(item);
-              return (
-                <div
-                  key={item.product.id}
-                  className='flex justify-between items-center bg-white p-3 rounded shadow'
-                >
-                  <div>
-                    <span className='font-semibold'>{item.product.name}</span>
-                    <br />
-                    <span className='text-sm text-gray-600'>
-                      {item.product.price}원 x {item.quantity}
-                      {appliedDiscount > 0 && (
-                        <span className='text-green-600 ml-1'>
-                          ({(appliedDiscount * 100).toFixed(0)}% 할인 적용)
-                        </span>
-                      )}
-                    </span>
-                  </div>
-                  <div>
-                    <button
-                      onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
-                      className='bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400'
-                    >
-                      -
-                    </button>
-                    <button
-                      onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
-                      className='bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400'
-                    >
-                      +
-                    </button>
-                    <button
-                      onClick={() => removeFromCart(item.product.id)}
-                      className='bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600'
-                    >
-                      삭제
-                    </button>
-                  </div>
-                </div>
-              );
-            })}
+            {cart.map((item) => (
+              <CartItemCard
+                item={item}
+                removeFromCart={removeFromCart}
+                updateQuantity={updateQuantity}
+              />
+            ))}
           </div>
 
           <div className='mt-6 bg-white p-4 rounded shadow'>
